+++ CHANGELOG.md
@@ -4,6 +4,7 @@ BREAKING CHANGES:
 
   * TransmissionClient constructor changed
   * option --host renamed to --transmission-host
+  * option --dest renamed to --download-torrents-dir
 
 
 FEATURES:
@@ -13,8 +14,8 @@ FEATURES:
 IMPROVEMENTS:
   * core: Config, TransmissionClient and WeburgClient fully tested
   * core: extracted WeburgClient class and fully tested
-  * command/download-weburg: command rewrited, fully tested 
-  * command/download-weburg: add logging of corrupted movie info 
+  * command/download-weburg: command rewrited, fully tested
+  * command/download-weburg: add logging of corrupted movie info
   * command/list-torrents: partially tested
   * command/remove-duplicates: partially tested